AbstractWaterfall: base class for waterfalls
Reduces code duplication, simplifies callers using both through common
interface, and generally simplifies maintenance. Functionally mostly
identical to the two old waterfall implementations.
