+++ GLWaterfall.cpp
@@ -1,7 +1,9 @@
- /* -*- c++ -*- */
+/* -*- c++ -*- */
 /* + + +   This Software is released under the "Simplified BSD License"  + + +
  * Copyright 2010 Moe Wheatley. All rights reserved.
  * Copyright 2011-2013 Alexandru Csete OZ9AEC
+ * Copyright 2018 Gonzalo José Carracedo Carballal - Minimal modifications for integration
+ * Copyright 2024 Sultan Qasim Khan - Abstract class for OpenGL and QPainter waterfalls
  *
  * Redistribution and use in source and binary forms, with or without modification, are
  * permitted provided that the following conditions are met:
@@ -39,6 +41,11 @@
 #include <QApplication>
 #include <cstring>
 
+// Apple deprecated OpenGL, I don't need to be warned, but for now it still works
+#ifdef __APPLE__
+#define GL_SILENCE_DEPRECATION
+#endif
+
 #include "SuWidgetsHelpers.h"
 #include "GLWaterfall.h"
 #include "gradient.h"
@@ -53,8 +60,8 @@
 #define GL_WATERFALL_MIN_BULK_TRANSFER  10
 
 struct vertex {
-    float vertex_coords[3];
-    float texture_coords[2];
+  float vertex_coords[3];
+  float texture_coords[2];
 };
 
 static const vertex vertices[] =
@@ -68,46 +75,46 @@ static const vertex vertices[] =
 static const unsigned short vertex_indices[] = {0, 1, 2, 2, 3, 0};
 
 static const char *wfVertexShader   = "                                        \
-    attribute vec3 vertex_coords;                                              \
-    attribute vec2 texture_coords;                                             \
-    varying   vec2 f_texture_coords;                                           \
-    uniform   mat4 ortho;                                                      \
-                                                                               \
-    void                                                                       \
-    main(void)                                                                 \
-    {                                                                          \
-       gl_Position      = ortho * vec4(vertex_coords, 1.0);                    \
-       f_texture_coords = texture_coords;                                      \
-    }                                                                          \
-";
+                                       attribute vec3 vertex_coords;                                              \
+                                       attribute vec2 texture_coords;                                             \
+                                       varying   vec2 f_texture_coords;                                           \
+                                       uniform   mat4 ortho;                                                      \
+                                       \
+                                       void                                                                       \
+                                       main(void)                                                                 \
+{                                                                          \
+  gl_Position      = ortho * vec4(vertex_coords, 1.0);                    \
+    f_texture_coords = texture_coords;                                      \
+}                                                                          \
+  ";
 
 static const char *wfFragmentShader = "                                        \
-    varying vec2      f_texture_coords;                                        \
-    uniform sampler2D m_texture;                                               \
-    uniform sampler2D m_palette;                                               \
-    uniform float     t;                                                       \
-    uniform float     x0;                                                      \
-    uniform float     m;                                                       \
-    uniform float     c_x0;                                                    \
-    uniform float     c_m;                                                     \
-                                                                               \
-    void                                                                       \
-    main(void)                                                                 \
-    {                                                                          \
-      float x = f_texture_coords.x * c_m + c_x0;                               \
-      float y = f_texture_coords.y + t - floor(f_texture_coords.y + t);        \
-      vec2 coord = vec2(x, y);                                                 \
-                                                                               \
-      vec4 psd = texture2D(m_texture, coord);                                  \
-      float paletteIndex = (psd.r - x0) / m;                                   \
-      vec4 palColor       = texture2D(m_palette, vec2(paletteIndex, 0));       \
-                                                                               \
-      gl_FragColor = palColor;                                                 \
-    }                                                                          \
-";
+                                       varying vec2      f_texture_coords;                                        \
+                                       uniform sampler2D m_texture;                                               \
+                                       uniform sampler2D m_palette;                                               \
+                                       uniform float     t;                                                       \
+                                       uniform float     x0;                                                      \
+                                       uniform float     m;                                                       \
+                                       uniform float     c_x0;                                                    \
+                                       uniform float     c_m;                                                     \
+                                       \
+                                       void                                                                       \
+                                       main(void)                                                                 \
+{                                                                          \
+  float x = f_texture_coords.x * c_m + c_x0;                               \
+    float y = f_texture_coords.y + t - floor(f_texture_coords.y + t);        \
+    vec2 coord = vec2(x, y);                                                 \
+    \
+    vec4 psd = texture2D(m_texture, coord);                                  \
+    float paletteIndex = (psd.r - x0) / m;                                   \
+    vec4 palColor       = texture2D(m_palette, vec2(paletteIndex, 0));       \
+    \
+    gl_FragColor = palColor;                                                 \
+}                                                                          \
+  ";
 
 ///////////////////////////// GLLine //////////////////////////////////////////
-void
+  void
 GLLine::normalize(void)
 {
   int i;
@@ -120,7 +127,7 @@ GLLine::normalize(void)
 
 }
 
-void
+  void
 GLLine::rescaleMean(void)
 {
   float *data = this->data();
@@ -141,7 +148,7 @@ GLLine::rescaleMean(void)
   }
 }
 
-void
+  void
 GLLine::rescaleMax(void)
 {
   float *data = this->data();
@@ -162,7 +169,7 @@ GLLine::rescaleMax(void)
   }
 }
 
-void
+  void
 GLLine::assignMean(const float *values)
 {
   float *data = this->data();
@@ -173,7 +180,7 @@ GLLine::assignMean(const float *values)
   rescaleMean();
 }
 
-void
+  void
 GLLine::assignMax(const float *values)
 {
   float *data = this->data();
@@ -185,7 +192,7 @@ GLLine::assignMax(const float *values)
 }
 
 
-void
+  void
 GLLine::reduceMean(const float *values, int length)
 {
   int res       = resolution();
@@ -208,7 +215,7 @@ GLLine::reduceMean(const float *values, int length)
   }
 }
 
-void
+  void
 GLLine::reduceMax(const float *values, int length)
 {
   int res       = resolution();
@@ -261,7 +268,7 @@ GLWaterfallOpenGLContext::~GLWaterfallOpenGLContext(void)
   delete m_palette;
 }
 
-void
+  void
 GLWaterfallOpenGLContext::initialize(void)
 {
   GLint texSize;
@@ -270,13 +277,13 @@ GLWaterfallOpenGLContext::initialize(void)
 
   for (int i = 0; i < 256; ++i) {
     firstPal.setPixel(
-          i,
-          0,
-          qRgba(
-            255 * wf_gradient[i][0],
-            255 * wf_gradient[i][1],
-            255 * wf_gradient[i][2],
-            255));
+        i,
+        0,
+        qRgba(
+          255 * wf_gradient[i][0],
+          255 * wf_gradient[i][1],
+          255 * wf_gradient[i][2],
+          255));
   }
 
   // Retrieve limits
@@ -334,7 +341,7 @@ GLWaterfallOpenGLContext::initialize(void)
   m_program.bind();
 }
 
-void
+  void
 GLWaterfallOpenGLContext::resetWaterfall(void)
 {
   GLLine nullLine;
@@ -350,43 +357,43 @@ GLWaterfallOpenGLContext::resetWaterfall(void)
   m_waterfall->setMinificationFilter(QOpenGLTexture::Linear);
   m_waterfall->setMagnificationFilter(QOpenGLTexture::Linear);
   m_waterfall->allocateStorage(
-        QOpenGLTexture::PixelFormat::Red,
-        QOpenGLTexture::PixelType::UInt32);
+      QOpenGLTexture::PixelFormat::Red,
+      QOpenGLTexture::PixelType::UInt32);
   m_waterfall->create();
   m_waterfall->bind(0);
 
   // Clear waterfall
   for (int i = 0; i < m_rowCount; ++i)
     glTexSubImage2D(
-          GL_TEXTURE_2D,
-          0,
-          0,
-          i,
-          nullLine.allocation(),
-          1,
-          GL_RED,
-          GL_FLOAT,
-          nullLine.data());
+        GL_TEXTURE_2D,
+        0,
+        0,
+        i,
+        nullLine.allocation(),
+        1,
+        GL_RED,
+        GL_FLOAT,
+        nullLine.data());
 
   m_row = 0;
 }
 
-void
+  void
 GLWaterfallOpenGLContext::flushPalette(void)
 {
   glTexSubImage2D(
-        GL_TEXTURE_2D,
-        0,
-        0,
-        0,
-        256,
-        1,
-        GL_RGBA,
-        GL_UNSIGNED_BYTE,
-        m_paletBuf.data());
+      GL_TEXTURE_2D,
+      0,
+      0,
+      0,
+      256,
+      1,
+      GL_RGBA,
+      GL_UNSIGNED_BYTE,
+      m_paletBuf.data());
 }
 
-void
+  void
 GLWaterfallOpenGLContext::disposeLastLine(void)
 {
   if (!m_history.empty()) {
@@ -404,7 +411,7 @@ GLWaterfallOpenGLContext::disposeLastLine(void)
   }
 }
 
-void
+  void
 GLWaterfallOpenGLContext::flushOneLine(void)
 {
   GLLine &line = m_history.back();
@@ -412,15 +419,15 @@ GLWaterfallOpenGLContext::flushOneLine(void)
 
   if (m_rowSize == line.resolution()) {
     glTexSubImage2D(
-          GL_TEXTURE_2D,
-          0,
-          0,
-          row,
-          line.allocation(),
-          1,
-          GL_RED,
-          GL_FLOAT,
-          line.data());
+        GL_TEXTURE_2D,
+        0,
+        0,
+        row,
+        line.allocation(),
+        1,
+        GL_RED,
+        GL_FLOAT,
+        line.data());
     this->disposeLastLine();
     m_row = (m_row + 1) % m_rowCount;
   } else {
@@ -429,7 +436,7 @@ GLWaterfallOpenGLContext::flushOneLine(void)
   }
 }
 
-void
+  void
 GLWaterfallOpenGLContext::flushLinesBulk(void)
 {
   int maxRows = m_rowCount - (m_row % m_rowCount);
@@ -448,9 +455,9 @@ GLWaterfallOpenGLContext::flushLinesBulk(void)
     }
 
     memcpy(
-          bulkData.data() + (maxRows - i - 1) * alloc,
-          line.data(),
-          alloc * sizeof(float));
+        bulkData.data() + (maxRows - i - 1) * alloc,
+        line.data(),
+        alloc * sizeof(float));
     this->disposeLastLine();
 
     ++count;
@@ -458,20 +465,20 @@ GLWaterfallOpenGLContext::flushLinesBulk(void)
 
   if (count > 0) {
     glTexSubImage2D(
-          GL_TEXTURE_2D,
-          0,
-          0,
-          maxRows - count,
-          alloc,
-          count,
-          GL_RED,
-          GL_FLOAT,
-          bulkData.data() + (maxRows - count) * alloc);
+        GL_TEXTURE_2D,
+        0,
+        0,
+        maxRows - count,
+        alloc,
+        count,
+        GL_RED,
+        GL_FLOAT,
+        bulkData.data() + (maxRows - count) * alloc);
     m_row = (m_row + count) % m_rowCount;
   }
 }
 
-void
+  void
 GLWaterfallOpenGLContext::flushLines(void)
 {
   while (!m_history.empty()) {
@@ -482,13 +489,13 @@ GLWaterfallOpenGLContext::flushLines(void)
   }
 }
 
-void
+  void
 GLWaterfallOpenGLContext::flushLinePool(void)
 {
   m_pool.clear();
 }
 
-void
+  void
 GLWaterfallOpenGLContext::setPalette(const QColor *table)
 {
   uint8_t *asBytes = reinterpret_cast<uint8_t *>(m_paletBuf.data());
@@ -503,7 +510,7 @@ GLWaterfallOpenGLContext::setPalette(const QColor *table)
   m_updatePalette = true;
 }
 
-void
+  void
 GLWaterfallOpenGLContext::pushFFTData(
     const float *__restrict__ fftData,
     int size)
@@ -551,7 +558,7 @@ GLWaterfallOpenGLContext::pushFFTData(
   }
 }
 
-void
+  void
 GLWaterfallOpenGLContext::averageFFTData(const float *fftData, int size)
 {
   int i;
@@ -570,21 +577,21 @@ GLWaterfallOpenGLContext::averageFFTData(const float *fftData, int size)
   }
 }
 
-void
+  void
 GLWaterfallOpenGLContext::commitFFTData(void)
 {
   pushFFTData(m_accum.data(), m_accum.size());
   m_firstAccum = true;
 }
 
-void
+  void
 GLWaterfallOpenGLContext::setDynamicRange(float mindB, float maxdB)
 {
   this->m  = (maxdB - mindB) / GL_WATERFALL_TEX_DR;
   this->x0 = (mindB - GL_WATERFALL_TEX_MIN_DB) / GL_WATERFALL_TEX_DR;
 }
 
-void
+  void
 GLWaterfallOpenGLContext::finalize()
 {
   if (m_vao.isCreated())
@@ -599,7 +606,7 @@ GLWaterfallOpenGLContext::finalize()
     m_palette->destroy();
 }
 
-void
+  void
 GLWaterfallOpenGLContext::recalcGeometric(int width, int height, float zoom)
 {
   int d = static_cast<int>(floor(log2(m_rowSize / (width * zoom))));
@@ -616,7 +623,7 @@ GLWaterfallOpenGLContext::recalcGeometric(int width, int height, float zoom)
   m_zoom   = zoom;
 }
 
-void
+  void
 GLWaterfallOpenGLContext::render(
     int x,
     int y,
@@ -647,18 +654,18 @@ GLWaterfallOpenGLContext::render(
   ortho.scale(zoom, 1);
 
   m_program.setAttributeBuffer(
-        "vertex_coords",
-        GL_FLOAT,
-        0,
-        3,
-        sizeof(vertex));
+      "vertex_coords",
+      GL_FLOAT,
+      0,
+      3,
+      sizeof(vertex));
 
   m_program.setAttributeBuffer(
-        "texture_coords",
-        GL_FLOAT,
-        sizeof(vertices[0].vertex_coords),
-        2,
-        sizeof(vertex));
+      "texture_coords",
+      GL_FLOAT,
+      sizeof(vertices[0].vertex_coords),
+      2,
+      sizeof(vertex));
 
   m_program.enableAttributeArray("vertex_coords");
   m_program.enableAttributeArray("texture_coords");
@@ -702,112 +709,9 @@ GLWaterfallOpenGLContext::render(
   glPopAttrib();
 }
 
-/////////////////////////////// GLWaterfall ///////////////////////////////////
-
-#define STATUS_TIP \
-    "Click, drag or scroll on spectrum to tune. " \
-    "Drag and scroll X and Y axes for pan and zoom. " \
-    "Drag filter edges to adjust filter."
-
-
-void
-GLWaterfall::initLayout(void)
-{
-  setSizePolicy(QSizePolicy::Expanding, QSizePolicy::Expanding);
-  setFocusPolicy(Qt::StrongFocus);
-  setAttribute(Qt::WA_PaintOnScreen,false);
-  setAutoFillBackground(false);
-  setAttribute(Qt::WA_OpaquePaintEvent, false);
-  setAttribute(Qt::WA_NoSystemBackground, true);
-  setMouseTracking(true);
-  setTooltipsEnabled(false);
-  setStatusTip(tr(STATUS_TIP));
-}
-
-void
-GLWaterfall::initDefaults(void)
-{
-  m_PeakHoldActive = false;
-  m_PeakHoldValid = false;
-
-  m_FftCenter = 0;
-  m_CenterFreq = 144500000;
-  m_DemodCenterFreq = 144500000;
-  m_DemodHiCutFreq = 5000;
-  m_DemodLowCutFreq = -5000;
-
-  m_FLowCmin = -25000;
-  m_FLowCmax = -1000;
-  m_FHiCmin = 1000;
-  m_FHiCmax = 25000;
-  m_symetric = true;
-
-  m_ClickResolution = 100;
-  m_FilterClickResolution = 100;
-  m_CursorCaptureDelta = CUR_CUT_DELTA;
-
-  m_FilterBoxEnabled = true;
-  m_CenterLineEnabled = true;
-  m_BookmarksEnabled = true;
-  m_Locked = false;
-  m_freqDragLocked = false;
-  m_Span = 96000;
-  m_SampleFreq = 96000;
-
-  m_HorDivs = 12;
-  m_VerDivs = 6;
-  m_PandMaxdB = m_WfMaxdB = 0.f;
-  m_PandMindB = m_WfMindB = -150.f;
-
-  m_CumWheelDelta = 0;
-  m_FreqUnits = 1000000;
-  m_CursorCaptured = NOCAP;
-  m_Running = false;
-  m_DrawOverlay = true;
-  m_2DPixmap = QPixmap(0,0);
-  m_OverlayPixmap = QPixmap(0,0);
-  m_GLWaterfallImage = QImage();
-  m_Size = QSize(0,0);
-  m_GrabPosition = 0;
-  m_Percent2DScreen = 30;	//percent of screen used for 2D display
-  m_VdivDelta = 30;
-  m_HdivDelta = 70;
-
-  m_ZeroPoint = 0;
-  m_dBPerUnit = 1;
-
-  m_FreqDigits = 3;
-
-  m_Peaks = QMap<int,int>();
-  setPeakDetection(false, 2);
-  m_PeakHoldValid = false;
-
-  setFftPlotColor(QColor(0xFF,0xFF,0xFF,0xFF));
-  setFftBgColor(QColor(PLOTTER_BGD_COLOR));
-  setFftAxesColor(QColor(PLOTTER_GRID_COLOR));
-  setFilterBoxColor(QColor(PLOTTER_FILTER_BOX_COLOR));
-  setTimeStampColor(QColor(0xFF,0xFF,0xFF,0xFF));
-
-  setFftFill(false);
-
-  // always update waterfall
-  tlast_wf_ms = 0;
-  msec_per_wfline = 0;
-  wf_span = 0;
-  fft_rate = 15;
-
-  m_fftData = nullptr;
-  m_wfData  = nullptr;
-  m_fftDataSize = 0;
-
-  m_infoTextColor = QColor(0xFF,0xFF,0xFF,0xFF);
-}
-
 ///////////////////////////// GLWaterfall ////////////////////////////////////////
-GLWaterfall::GLWaterfall(QWidget *parent) : QOpenGLWidget(parent)
+GLWaterfall::GLWaterfall(QWidget *parent) : AbstractWaterfall(parent)
 {
-  this->initLayout();
-  this->initDefaults();
 }
 
 GLWaterfall::~GLWaterfall()
@@ -817,325 +721,10 @@ GLWaterfall::~GLWaterfall()
   doneCurrent();
 }
 
-QSize
-GLWaterfall::minimumSizeHint() const
+  void
+GLWaterfall::clearWaterfall()
 {
-  return QSize(50, 50);
-}
-
-QSize
-GLWaterfall::sizeHint() const
-{
-  return QSize(180, 180);
-}
-
-void
-GLWaterfall::mouseMoveEvent(QMouseEvent *event)
-{
-  QPoint pt = event->pos();
-
-  /* mouse enter / mouse leave events */
-  if (m_OverlayPixmap.rect().contains(pt)) {
-    //is in Overlay bitmap region
-    if (event->buttons() == Qt::NoButton) {
-      bool onTag = false;
-      if (pt.y() < 15 * 10) { // FIXME
-        for (int i = 0; i < m_BookmarkTags.size() && !onTag; i++) {
-          if (m_BookmarkTags[i].first.contains(event->pos()))
-            onTag = true;
-        }
-      }
-      // if no mouse button monitor grab regions and change cursor icon
-      if (onTag) {
-        setCursor(QCursor(Qt::PointingHandCursor));
-        m_CursorCaptured = BOOKMARK;
-      } else if (isPointCloseTo(pt.x(), m_DemodFreqX, m_CursorCaptureDelta)) {
-        // in move demod box center frequency region
-        if (m_CursorCaptured != CENTER)
-          setCursor(QCursor(Qt::SizeHorCursor));
-        m_CursorCaptured = CENTER;
-        if (m_TooltipsEnabled)
-          QToolTip::showText(
-#if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
-                event->globalPosition().toPoint(),
-#else
-                event->globalPos(),
-#endif
-                QString("Demod: %1 kHz").arg(
-                  m_DemodCenterFreq / 1.e3f, 0, 'f', 3),
-                this);
-      } else if (isPointCloseTo(
-                   pt.x(),
-                   m_DemodHiCutFreqX,
-                   m_CursorCaptureDelta)) {
-        // in move demod hicut region
-        if (m_CursorCaptured != RIGHT)
-          setCursor(QCursor(Qt::SizeFDiagCursor));
-        m_CursorCaptured = RIGHT;
-        if (m_TooltipsEnabled)
-          QToolTip::showText(
-#if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
-                event->globalPosition().toPoint(),
-#else
-                event->globalPos(),
-#endif
-                QString("High cut: %1 Hz").arg(m_DemodHiCutFreq),
-                this);
-      } else if (isPointCloseTo(
-                   pt.x(),
-                   m_DemodLowCutFreqX,
-                   m_CursorCaptureDelta)) {
-        // in move demod lowcut region
-        if (m_CursorCaptured != LEFT)
-          setCursor(QCursor(Qt::SizeBDiagCursor));
-        m_CursorCaptured = LEFT;
-        if (m_TooltipsEnabled)
-          QToolTip::showText(
-#if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
-                event->globalPosition().toPoint(),
-#else
-                event->globalPos(),
-#endif
-                QString("Low cut: %1 Hz").arg(m_DemodLowCutFreq),
-                this);
-      } else if (isPointCloseTo(
-                   pt.x(),
-                   m_YAxisWidth / 2,
-                   m_YAxisWidth / 2)) {
-        if (m_CursorCaptured != YAXIS)
-          setCursor(QCursor(Qt::OpenHandCursor));
-        m_CursorCaptured = YAXIS;
-        if (m_TooltipsEnabled)
-          QToolTip::hideText();
-      } else if (isPointCloseTo(
-                   pt.y(),
-                   m_XAxisYCenter,
-                   m_CursorCaptureDelta + 5)) {
-        if (m_CursorCaptured != XAXIS)
-          setCursor(QCursor(Qt::OpenHandCursor));
-        m_CursorCaptured = XAXIS;
-        if (m_TooltipsEnabled)
-          QToolTip::hideText();
-      } else {	//if not near any grab boundaries
-        if (m_CursorCaptured != NOCAP) {
-          setCursor(QCursor(Qt::ArrowCursor));
-          m_CursorCaptured = NOCAP;
-        }
-        if (m_TooltipsEnabled)
-          QToolTip::showText(
-#if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
-                event->globalPosition().toPoint(),
-#else
-                event->globalPos(),
-#endif
-                QString("F: %1 kHz").arg(freqFromX(pt.x())/1.e3f, 0, 'f', 3),
-                this);
-      }
-      m_GrabPosition = 0;
-    }
-  } else {
-    // not in Overlay region
-    if (event->buttons() == Qt::NoButton) {
-      if (NOCAP != m_CursorCaptured)
-        setCursor(QCursor(Qt::ArrowCursor));
-
-      m_CursorCaptured = NOCAP;
-      m_GrabPosition = 0;
-    }
-    if (m_TooltipsEnabled) {
-      QDateTime tt;
-      tt.setMSecsSinceEpoch(msecFromY(pt.y()));
-
-      QToolTip::showText(
-#if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
-            event->globalPosition().toPoint(),
-#else
-            event->globalPos(),
-#endif
-            QString("%1\n%2 kHz")
-              .arg(tt.toString("yyyy.MM.dd hh:mm:ss.zzz"))
-              .arg(freqFromX(pt.x())/1.e3f, 0, 'f', 3),
-            this);
-    }
-  }
-  // process mouse moves while in cursor capture modes
-  if (YAXIS == m_CursorCaptured) {
-    if (event->buttons() & Qt::LeftButton) {
-      setCursor(QCursor(Qt::ClosedHandCursor));
-      // move Y scale up/down
-      float delta_px = m_Yzero - pt.y();
-      float delta_db =
-          delta_px * fabs(m_PandMindB - m_PandMaxdB) / (float)m_OverlayPixmap.height();
-      m_PandMindB -= delta_db;
-      m_PandMaxdB -= delta_db;
-      if (out_of_range(m_PandMindB, m_PandMaxdB)) {
-        m_PandMindB += delta_db;
-        m_PandMaxdB += delta_db;
-      } else {
-        emit pandapterRangeChanged(m_PandMindB, m_PandMaxdB);
-        updateOverlay();
-        m_PeakHoldValid = false;
-        m_Yzero = pt.y();
-      }
-    }
-  } else if (XAXIS == m_CursorCaptured) {
-#if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
-    if (event->buttons() & (Qt::LeftButton | Qt::MiddleButton))
-#else
-    if (event->buttons() & (Qt::LeftButton | Qt::MidButton))
-#endif // QT_VERSION
-    {
-      setCursor(QCursor(Qt::ClosedHandCursor));
-      // pan viewable range or move center frequency
-      int delta_px = m_Xzero - pt.x();
-      qint64 delta_hz = delta_px * m_Span / m_OverlayPixmap.width();
-      if (event->buttons() & m_freqDragBtn) {
-        if (!m_Locked && !m_freqDragLocked) {
-          qint64 centerFreq = boundCenterFreq(m_CenterFreq + delta_hz);
-          delta_hz = centerFreq - m_CenterFreq;
-
-          m_CenterFreq += delta_hz;
-          m_DemodCenterFreq += delta_hz;
-
-          ////////////// TODO: Edit this to fake spectrum scroll /////////
-          ////////////// DONE: Something like this:
-          ///
-          m_tentativeCenterFreq += delta_hz;
-
-          if (delta_hz != 0)
-            emit newCenterFreq(m_CenterFreq);
-        }
-      } else {
-        setFftCenterFreq(m_FftCenter + delta_hz);
-      }
-
-      updateOverlay();
-      m_PeakHoldValid = false;
-      m_Xzero = pt.x();
-    }
-  } else if (LEFT == m_CursorCaptured) {
-    // moving in demod lowcut region
-    if (event->buttons() & (Qt::LeftButton | Qt::RightButton)) {
-      // moving in demod lowcut region with left button held
-      if (m_GrabPosition != 0) {
-        m_DemodLowCutFreq = freqFromX(pt.x() - m_GrabPosition ) - m_DemodCenterFreq;
-        m_DemodLowCutFreq = roundFreq(m_DemodLowCutFreq, m_FilterClickResolution);
-
-        if (m_symetric && (event->buttons() & Qt::LeftButton)) {
-          m_DemodHiCutFreq = -m_DemodLowCutFreq;
-        }
-        clampDemodParameters();
-
-        emit newFilterFreq(m_DemodLowCutFreq, m_DemodHiCutFreq);
-        updateOverlay();
-      } else {
-        // save initial grab postion from m_DemodFreqX
-        m_GrabPosition = pt.x()-m_DemodLowCutFreqX;
-      }
-    } else if (event->buttons() & ~Qt::NoButton) {
-      setCursor(QCursor(Qt::ArrowCursor));
-      m_CursorCaptured = NOCAP;
-    }
-  } else if (RIGHT == m_CursorCaptured) {
-    // moving in demod highcut region
-    if (event->buttons() & (Qt::LeftButton | Qt::RightButton)) {
-      // moving in demod highcut region with right button held
-      if (m_GrabPosition != 0) {
-        m_DemodHiCutFreq = freqFromX( pt.x()-m_GrabPosition ) - m_DemodCenterFreq;
-        m_DemodHiCutFreq = roundFreq(m_DemodHiCutFreq, m_FilterClickResolution);
-
-        if (m_symetric && (event->buttons() & Qt::LeftButton)) {
-          m_DemodLowCutFreq = -m_DemodHiCutFreq;
-        }
-        clampDemodParameters();
-
-        emit newFilterFreq(m_DemodLowCutFreq, m_DemodHiCutFreq);
-        updateOverlay();
-      } else {
-        // save initial grab postion from m_DemodFreqX
-        m_GrabPosition = pt.x() - m_DemodHiCutFreqX;
-      }
-    } else if (event->buttons() & ~Qt::NoButton) {
-      setCursor(QCursor(Qt::ArrowCursor));
-      m_CursorCaptured = NOCAP;
-    }
-  } else if (CENTER == m_CursorCaptured) {
-    // moving inbetween demod lowcut and highcut region
-    if (event->buttons() & Qt::LeftButton) {   // moving inbetween demod lowcut and highcut region with left button held
-      if (m_GrabPosition != 0) {
-        if (!m_Locked) {
-          m_DemodCenterFreq = roundFreq(
-                freqFromX(pt.x() - m_GrabPosition),
-                m_ClickResolution);
-          emit newDemodFreq(
-                m_DemodCenterFreq,
-                m_DemodCenterFreq - m_CenterFreq);
-          updateOverlay();
-          m_PeakHoldValid = false;
-        }
-      } else {
-        // save initial grab postion from m_DemodFreqX
-        m_GrabPosition = pt.x() - m_DemodFreqX;
-      }
-    } else if (event->buttons() & ~Qt::NoButton) {
-      setCursor(QCursor(Qt::ArrowCursor));
-      m_CursorCaptured = NOCAP;
-    }
-  } else {
-    // cursor not captured
-    m_GrabPosition = 0;
-  }
-
-  if (!this->rect().contains(pt)) {
-    if (NOCAP != m_CursorCaptured)
-      setCursor(QCursor(Qt::ArrowCursor));
-    m_CursorCaptured = NOCAP;
-  }
-}
-
-
-int
-GLWaterfall::getNearestPeak(QPoint pt)
-{
-  QMap<int, int>::const_iterator i =
-      m_Peaks.lowerBound(pt.x() - PEAK_CLICK_MAX_H_DISTANCE);
-  QMap<int, int>::const_iterator upperBound =
-      m_Peaks.upperBound(pt.x() + PEAK_CLICK_MAX_H_DISTANCE);
-  float   dist = 1.0e10;
-  int     best = -1;
-
-  for ( ; i != upperBound; i++) {
-    int x = i.key();
-    int y = i.value();
-
-    if (abs(y - pt.y()) > PEAK_CLICK_MAX_V_DISTANCE)
-      continue;
-
-    float d = powf(y - pt.y(), 2) + powf(x - pt.x(), 2);
-    if (d < dist) {
-      dist = d;
-      best = x;
-    }
-  }
-
-  return best;
-}
-
-/** Set waterfall span in milliseconds */
-void
-GLWaterfall::setWaterfallSpan(quint64 span_ms)
-{
-  wf_span = span_ms;
-  if (m_GLWaterfallImage.height() > 0)
-    msec_per_wfline = wf_span / (m_GLWaterfallImage.height() *
-            screen()->devicePixelRatio());
-  clearGLWaterfall();
-}
-
-void
-GLWaterfall::clearGLWaterfall()
-{
-  m_GLWaterfallImage.fill(Qt::black);
+  m_WaterfallImage.fill(Qt::black);
 }
 
 /**
@@ -1146,330 +735,44 @@ GLWaterfall::clearGLWaterfall()
  * We assume that frequency strings are up to date
  */
 bool
-GLWaterfall::saveGLWaterfall(const QString &) const
+GLWaterfall::saveWaterfall(const QString &) const
 {
   return false;
 }
 
-/** Get waterfall time resolution in milleconds / line. */
-double
-GLWaterfall::getWfTimeRes(void)
-{
-  if (msec_per_wfline)
-    return msec_per_wfline;
-  else
-    return 1000.0 / fft_rate;
-}
-
-void
-GLWaterfall::setFftRate(int rate_hz)
-{
-  fft_rate = rate_hz;
-  clearGLWaterfall();
-}
-
-// Called when a mouse button is pressed
-void
-GLWaterfall::mousePressEvent(QMouseEvent * event)
-{
-  QPoint pt = event->pos();
-
-  if (NOCAP == m_CursorCaptured) {
-    if (isPointCloseTo(pt.x(), m_DemodFreqX, m_CursorCaptureDelta)) {
-      // move demod box center frequency region
-      m_CursorCaptured = CENTER;
-      m_GrabPosition = pt.x() - m_DemodFreqX;
-    } else if (isPointCloseTo(
-                 pt.x(),
-                 m_DemodLowCutFreqX,
-                 m_CursorCaptureDelta)) {
-      // filter low cut
-      m_CursorCaptured = LEFT;
-      m_GrabPosition = pt.x() - m_DemodLowCutFreqX;
-    } else if (isPointCloseTo(
-                 pt.x(),
-                 m_DemodHiCutFreqX,
-                 m_CursorCaptureDelta)) {
-      // filter high cut
-      m_CursorCaptured = RIGHT;
-      m_GrabPosition = pt.x() - m_DemodHiCutFreqX;
-    } else {
-      if (event->buttons() == Qt::LeftButton) {
-        if (!m_Locked) {
-          int  best = -1;
-
-          if (m_PeakDetection > 0)
-            best = getNearestPeak(pt);
-          if (best != -1)
-            m_DemodCenterFreq = freqFromX(best);
-          else
-            m_DemodCenterFreq = roundFreq(freqFromX(pt.x()), m_ClickResolution);
-
-          // if cursor not captured set demod frequency and start demod box capture
-          emit newDemodFreq(m_DemodCenterFreq, m_DemodCenterFreq - m_CenterFreq);
-
-          // save initial grab postion from m_DemodFreqX
-          // setCursor(QCursor(Qt::CrossCursor));
-          m_CursorCaptured = CENTER;
-          m_GrabPosition = 1;
-          updateOverlay();
-        }
-      }
-#if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
-      else if (event->buttons() == Qt::MiddleButton)
-#else
-      else if (event->buttons() == Qt::MidButton)
-#endif // QT_VERSION
-      {
-        if (!m_Locked && !m_freqDragLocked) {
-          // set center freq
-          m_CenterFreq
-              = boundCenterFreq(roundFreq(freqFromX(pt.x()), m_ClickResolution));
-          m_DemodCenterFreq = m_CenterFreq;
-          emit newCenterFreq(m_CenterFreq);
-          emit newDemodFreq(m_DemodCenterFreq, m_DemodCenterFreq - m_CenterFreq);
-          updateOverlay();
-        }
-      } else if (event->buttons() == Qt::RightButton) {
-        // reset frequency zoom
-        resetHorizontalZoom();
-        updateOverlay();
-      }
-    }
-  } else {
-    if (m_CursorCaptured == YAXIS)
-      // get ready for moving Y axis
-      m_Yzero = pt.y();
-    else if (m_CursorCaptured == XAXIS) {
-      m_Xzero = pt.x();
-      if (event->buttons() == Qt::RightButton) {
-        // reset frequency zoom
-        resetHorizontalZoom();
-        updateOverlay();
-      }
-    } else if (m_CursorCaptured == BOOKMARK) {
-      if (!m_Locked) {
-        for (int i = 0; i < m_BookmarkTags.size(); i++) {
-          if (m_BookmarkTags[i].first.contains(event->pos())) {
-            BookmarkInfo info = m_BookmarkTags[i].second;
-
-            if (!info.modulation.isEmpty()) {
-              emit newModulation(info.modulation);
-            }
-
-            m_DemodCenterFreq = info.frequency;
-            emit newDemodFreq(m_DemodCenterFreq, m_DemodCenterFreq - m_CenterFreq);
-
-            if (info.bandwidth() != 0) {
-              emit newFilterFreq(info.lowFreqCut, info.highFreqCut);
-            }
-
-            break;
-          }
-        }
-      }
-    }
-  }
-}
-
-void
-GLWaterfall::mouseReleaseEvent(QMouseEvent * event)
-{
-  QPoint pt = event->pos();
-
-  if (!m_OverlayPixmap.rect().contains(pt)) {
-    // not in Overlay region
-    if (NOCAP != m_CursorCaptured)
-      setCursor(QCursor(Qt::ArrowCursor));
-
-    m_CursorCaptured = NOCAP;
-    m_GrabPosition = 0;
-  } else {
-    if (YAXIS == m_CursorCaptured) {
-      setCursor(QCursor(Qt::OpenHandCursor));
-      m_Yzero = -1;
-    } else if (XAXIS == m_CursorCaptured) {
-      setCursor(QCursor(Qt::OpenHandCursor));
-      m_Xzero = -1;
-    }
-  }
-}
-
-
-// Make a single zoom step on the X axis.
-void
-GLWaterfall::zoomStepX(float step, int x)
-{
-  // calculate new range shown on FFT
-  qreal new_range = qBound(10.0,
-                           (qreal)(m_Span) * step,
-                           (qreal)(m_SampleFreq) * 10.0);
-
-  // Frequency where event occured is kept fixed under mouse
-  qreal ratio = (qreal)x / (qreal)m_OverlayPixmap.width();
-  qreal fixed_hz = freqFromX(x);
-  qreal f_max = fixed_hz + (1.0 - ratio) * new_range;
-  qreal f_min = f_max - new_range;
-
-  qint64 fc = (qint64)(f_min + (f_max - f_min) / 2.0);
-
-  setFftCenterFreq(fc - m_CenterFreq);
-  setSpanFreq(new_range);
-
-  qreal factor = (qreal)m_SampleFreq / (qreal)m_Span;
-  emit newZoomLevel(factor);
-
-  m_PeakHoldValid = false;
-}
-
-// Zoom on X axis (absolute level)
-void
-GLWaterfall::zoomOnXAxis(float level)
-{
-  float current_level = (float)m_SampleFreq / (float)m_Span;
-
-  zoomStepX(current_level / level, xFromFreq(m_DemodCenterFreq));
-}
-
-// Called when a mouse wheel is turned
-void
-GLWaterfall::wheelEvent(QWheelEvent * event)
-{
-  QOpenGLWidget::wheelEvent(event);
-#if QT_VERSION >= QT_VERSION_CHECK(5, 14, 0)
-    QPointF pt = event->position();
-#else
-    QPointF pt = event->pos();
-#endif // QT_VERSION
-
-  // delta is in eigths of a degree, 15 degrees is one step
-  double numSteps = event->angleDelta().y() / (8.0 * 15.0);
-
-  if (m_CursorCaptured == YAXIS) {
-    // Vertical zoom. Wheel down: zoom out, wheel up: zoom in
-    // During zoom we try to keep the point (dB or kHz) under the cursor fixed
-    qreal zoom_fac = pow(0.9, numSteps);
-    qreal ratio = pt.y() / m_OverlayPixmap.height();
-    qreal db_range = m_PandMaxdB - m_PandMindB;
-    qreal y_range = m_OverlayPixmap.height();
-    qreal db_per_pix = db_range / y_range;
-    qreal fixed_db = m_PandMaxdB - pt.y() * db_per_pix;
-
-    db_range = qBound(
-          10.,
-          db_range * zoom_fac,
-          SCAST(qreal, FFT_MAX_DB - FFT_MIN_DB));
-    m_PandMaxdB = fixed_db + ratio * db_range;
-    if (m_PandMaxdB > FFT_MAX_DB)
-      m_PandMaxdB = FFT_MAX_DB;
-
-    m_PandMindB = m_PandMaxdB - db_range;
-    if (m_PandMindB < FFT_MIN_DB)
-      m_PandMindB = FFT_MIN_DB;
-
-    m_PeakHoldValid = false;
-
-    emit pandapterRangeChanged(m_PandMindB, m_PandMaxdB);
-  } else if (m_CursorCaptured == XAXIS) {
-    zoomStepX(pow(0.9, numSteps), pt.x());
-  } else if (event->modifiers() & Qt::ControlModifier) {
-    // filter width
-    m_DemodLowCutFreq -= numSteps * m_ClickResolution;
-    m_DemodHiCutFreq += numSteps * m_ClickResolution;
-    clampDemodParameters();
-    emit newFilterFreq(m_DemodLowCutFreq, m_DemodHiCutFreq);
-  } else if (event->modifiers() & Qt::ShiftModifier) {
-    if (!m_Locked) {
-      // filter shift
-      m_DemodLowCutFreq += numSteps * m_ClickResolution;
-      m_DemodHiCutFreq += numSteps * m_ClickResolution;
-      clampDemodParameters();
-      emit newFilterFreq(m_DemodLowCutFreq, m_DemodHiCutFreq);
-    }
-  } else {
-    if (!m_Locked) {
-      // small steps will be lost by roundFreq, let them accumulate
-      m_CumWheelDelta += event->angleDelta().y();
-      if (abs(m_CumWheelDelta) < 8*15)
-        return;
-      numSteps = m_CumWheelDelta / (8.0 * 15.0);
-
-      // inc/dec demod frequency
-      m_DemodCenterFreq += (numSteps * m_ClickResolution);
-      m_DemodCenterFreq = roundFreq(m_DemodCenterFreq, m_ClickResolution );
-      emit newDemodFreq(m_DemodCenterFreq, m_DemodCenterFreq-m_CenterFreq);
-    }
-  }
-
-  updateOverlay();
-  m_CumWheelDelta = 0;
-}
-
-// Called when screen size changes so must recalculate bitmaps
-void
-GLWaterfall::resizeEvent(QResizeEvent* ev)
-{
-  if (ev != nullptr)
-    QOpenGLWidget::resizeEvent(ev);
-
-  if (!size().isValid())
-    return;
-
-  if (m_Size != size()) {
-    // if changed, resize pixmaps to new screensize
-    int     fft_plot_height;
-
-    m_Size = size();
-    fft_plot_height = m_Percent2DScreen * m_Size.height() / 100;
-    m_OverlayPixmap = QPixmap(m_Size.width(), fft_plot_height);
-    m_OverlayPixmap.fill(Qt::black);
-    m_2DPixmap = QPixmap(m_Size.width(), fft_plot_height);
-    m_2DPixmap.fill(Qt::black);
-
-    int height = (100 - m_Percent2DScreen) * m_Size.height() / 100;
-    if (m_GLWaterfallImage.isNull()) {
-      m_GLWaterfallImage = QImage(
-            m_Size.width(),
-            height,
-            QImage::Format::Format_RGB32);
-      m_GLWaterfallImage.fill(Qt::black);
-    } else {
-      m_GLWaterfallImage = m_GLWaterfallImage.scaled(
-            m_Size.width(),
-            height,
-            Qt::IgnoreAspectRatio,
-            Qt::SmoothTransformation);
-    }
-
-    m_PeakHoldValid = false;
-
-    if (wf_span > 0)
-      msec_per_wfline = wf_span / (height * screen()->devicePixelRatio());
-  }
-
-  updateOverlay();
-}
-
-void
+  void
 GLWaterfall::initializeGL(void)
 {
   this->glCtx.initialize();
 
   connect(
-        this->context(),
-        SIGNAL(aboutToBeDestroyed()),
-        this,
-        SLOT(onContextBeingDestroyed()));
+      this->context(),
+      SIGNAL(aboutToBeDestroyed()),
+      this,
+      SLOT(onContextBeingDestroyed()));
 
   this->drawOverlay();
 }
 
-void
+  void
 GLWaterfall::onContextBeingDestroyed()
 {
   makeCurrent();
   this->glCtx.finalize();
   doneCurrent();
+}
+
+void GLWaterfall::setWaterfallRange(float min, float max)
+{
+  if (out_of_range(min, max))
+    return;
+
+  m_WfMindB = min;
+  m_WfMaxdB = max;
+
+  this->glCtx.setDynamicRange(min - m_gain, max - m_gain);
+
+  // no overlay change is necessary
 }
 
 //
@@ -1479,7 +782,7 @@ GLWaterfall::onContextBeingDestroyed()
 //        f0       f1
 //
 
-void
+  void
 GLWaterfall::paintGL(void)
 {
   int y = m_Percent2DScreen * m_Size.height() / 100;
@@ -1489,17 +792,17 @@ GLWaterfall::paintGL(void)
   int dpi_factor = screen()->devicePixelRatio();
 
   this->glCtx.render(0, y * dpi_factor, width() * dpi_factor,
-          height() * dpi_factor, left, right);
+      height() * dpi_factor, left, right);
 }
 
-void
+  void
 GLWaterfall::resizeGL(int, int)
 {
 }
 
 
 // Called by QT when screen needs to be redrawn
-void
+  void
 GLWaterfall::paintEvent(QPaintEvent *ev)
 {
   QOpenGLWidget::paintEvent(ev);
@@ -1528,14 +831,14 @@ GLWaterfall::paintEvent(QPaintEvent *ev)
         break;
 
       WFHelpers::drawChannelCutoff(
-            painter,
-            y,
-            x_fMin,
-            x_fMax,
-            x_fCenter,
-            p->markerColor,
-            p->cutOffColor,
-            !p->bandLike);
+          painter,
+          y,
+          x_fMin,
+          x_fMax,
+          x_fCenter,
+          p->markerColor,
+          p->cutOffColor,
+          !p->bandLike);
     }
   }
 
@@ -1547,137 +850,11 @@ GLWaterfall::paintEvent(QPaintEvent *ev)
 
   if (m_TimeStampsEnabled)
     paintTimeStamps(
-          painter,
-          QRect(2, y, this->width(), this->height()));
-}
-
-void
-GLWaterfall::paintTimeStamps(
-    QPainter &painter,
-    QRect const &where)
-{
-  QFontMetrics metrics(m_Font);
-  int y = where.y();
-  int textWidth;
-  int textHeight = metrics.height();
-  int items = 0;
-  int leftSpacing = 0;
-  int dpi_factor = screen()->devicePixelRatio();
-
-  auto it = m_TimeStamps.begin();
-
-  painter.setFont(m_Font);
-
-  y += m_TimeStampCounter / dpi_factor;
-
-  if (m_TimeStampMaxHeight < where.height())
-    m_TimeStampMaxHeight = where.height();
-
-  painter.setPen(QPen(m_TimeStampColor));
-
-#if QT_VERSION >= QT_VERSION_CHECK(5, 11, 0)
-  leftSpacing = metrics.horizontalAdvance("00:00:00.000");
-#else
-  leftSpacing = metrics.width("00:00:00.000");
-#endif // QT_VERSION_CHECK
-
-  while (y < m_TimeStampMaxHeight + textHeight && it != m_TimeStamps.end()) {
-    QString const &timeStampText =
-        m_TimeStampsUTC ? it->utcTimeStampText : it->timeStampText;
-#if QT_VERSION >= QT_VERSION_CHECK(5, 11, 0)
-    textWidth = metrics.horizontalAdvance(timeStampText);
-#else
-    textWidth = metrics.width(it->timeStampText);
-#endif // QT_VERSION_CHECK
-
-    if (it->marker) {
-      painter.drawText(
-            where.x() + where.width() - textWidth - 2,
-            y - 2,
-            timeStampText);
-      painter.drawLine(where.x() + leftSpacing, y, where.width() - 1, y);
-    } else {
-      painter.drawText(where.x(), y - 2, timeStampText);
-      painter.drawLine(where.x(), y, textWidth + where.x(), y);
-    }
-
-    y += it->counter / dpi_factor;
-    ++it;
-    ++items;
-  }
-
-  // TimeStamps from here could not be painted due to geometry restrictions.
-  // we silently discard them.
-
-  while (items < m_TimeStamps.size())
-    m_TimeStamps.removeLast();
-}
-
-void
-GLWaterfall::drawChannelBoxAndCutoff(
-    QPainter &painter,
-    int h,
-    qint64 fMin,
-    qint64 fMax,
-    qint64 fCenter,
-    QColor boxColor,
-    QColor markerColor,
-    QColor cutOffColor,
-    QString text,
-    QColor textColor)
-{
-
-  int x_fCenter = xFromFreq(fCenter);
-  int x_fMin = xFromFreq(fMin);
-  int x_fMax = xFromFreq(fMax);
-
-  WFHelpers::drawChannelBox(
-      painter,
-      h,
-      x_fMin,
-      x_fMax,
-      x_fCenter,
-      boxColor,
-      markerColor,
-      text,
-      textColor);
-
-  WFHelpers::drawChannelCutoff(
         painter,
-        h,
-        x_fMin,
-        x_fMax,
-        x_fCenter,
-        markerColor,
-        cutOffColor);
+        QRect(2, y, this->width(), this->height()));
 }
 
-void
-GLWaterfall::drawFilterBox(QPainter &painter, int h)
-{
-  m_DemodFreqX = xFromFreq(m_DemodCenterFreq);
-  m_DemodLowCutFreqX = xFromFreq(m_DemodCenterFreq + m_DemodLowCutFreq);
-  m_DemodHiCutFreqX = xFromFreq(m_DemodCenterFreq + m_DemodHiCutFreq);
-
-  drawChannelBoxAndCutoff(
-        painter,
-        h,
-        m_DemodCenterFreq + m_DemodLowCutFreq,
-        m_DemodCenterFreq + m_DemodHiCutFreq,
-        m_DemodCenterFreq,
-        m_FilterBoxColor,
-        QColor(PLOTTER_FILTER_LINE_COLOR),
-        m_TimeStampColor);
-}
-
-void
-GLWaterfall::drawFilterBox(GLDrawingContext &ctx)
-{
-  drawFilterBox(*ctx.painter, ctx.height);
-}
-
-
-void
+  void
 GLWaterfall::drawSpectrum(QPainter &painter, int forceHeight)
 {
   int     i, n;
@@ -1690,7 +867,7 @@ GLWaterfall::drawSpectrum(QPainter &painter, int forceHeight)
   // workaround for "fixed" line drawing since Qt 5
   // see http://stackoverflow.com/questions/16990326
 #if QT_VERSION >= 0x050000
-      painter.translate(0.5, 0.5);
+  painter.translate(0.5, 0.5);
 #endif
 
   // Do we have valid FFT data?
@@ -1699,22 +876,22 @@ GLWaterfall::drawSpectrum(QPainter &painter, int forceHeight)
 
   // get new scaled fft data
   getScreenIntegerFFTData(
-        h,
-        qMin(w, MAX_SCREENSIZE),
-        m_PandMaxdB,
-        m_PandMindB,
-        qBound(
-          -limit,
-          m_tentativeCenterFreq + m_FftCenter,
-          limit) - (qint64)m_Span/2,
-          qBound(
-            -limit,
-            m_tentativeCenterFreq + m_FftCenter,
-            limit) + (qint64)m_Span/2,
-          m_fftData,
-          m_fftbuf,
-          &xmin,
-          &xmax);
+      h,
+      qMin(w, MAX_SCREENSIZE),
+      m_PandMaxdB,
+      m_PandMindB,
+      qBound(
+        -limit,
+        m_tentativeCenterFreq + m_FftCenter,
+        limit) - (qint64)m_Span/2,
+      qBound(
+        -limit,
+        m_tentativeCenterFreq + m_FftCenter,
+        limit) + (qint64)m_Span/2,
+      m_fftData,
+      m_fftbuf,
+      &xmin,
+      &xmax);
 
   // draw the pandapter
   painter.setPen(m_FftColor);
@@ -1760,7 +937,7 @@ GLWaterfall::drawSpectrum(QPainter &painter, int forceHeight)
     for (i = 0; i < n; i++) {
       //m_PeakDetection times the std over the mean or better than current peak
       float d = (lastPeak == -1) ? (mean - m_PeakDetection * stdev) :
-                                   m_fftbuf[lastPeak + xmin];
+        m_fftbuf[lastPeak + xmin];
 
       if (m_fftbuf[i + xmin] < d)
         lastPeak=i;
@@ -1770,7 +947,7 @@ GLWaterfall::drawSpectrum(QPainter &painter, int forceHeight)
       {
         m_Peaks.insert(lastPeak + xmin, m_fftbuf[lastPeak + xmin]);
         painter.drawEllipse(lastPeak + xmin - 5,
-                            m_fftbuf[lastPeak + xmin] - 5, 10, 10);
+            m_fftbuf[lastPeak + xmin] - 5, 10, 10);
         lastPeak = -1;
       }
     }
@@ -1793,12 +970,13 @@ GLWaterfall::drawSpectrum(QPainter &painter, int forceHeight)
 }
 
 // Called to update spectrum data for displaying on the screen
-void
-GLWaterfall::draw()
+  void
+GLWaterfall::draw(bool everything)
 {
   int     w;
   int     h;
 
+  (void)everything; // suppress unused variable warning
 
   if (m_DrawOverlay) {
     drawOverlay();
@@ -1825,29 +1003,6 @@ GLWaterfall::draw()
 
 /**
  * Set new FFT data.
- * @param fftData Pointer to the new FFT data (same data for pandapter and waterfall).
- * @param size The FFT size.
- *
- * When FFT data is set using this method, the same data will be used for both the
- * pandapter and the waterfall.
- */
-void
-GLWaterfall::setNewFftData(
-    float *fftData,
-    int size,
-    QDateTime const &t,
-    bool looped)
-{
-  if (m_tentativeCenterFreq != 0) {
-    m_tentativeCenterFreq = 0;
-    m_DrawOverlay = true;
-  }
-
-  this->setNewFftData(fftData, fftData, size, t, looped);
-}
-
-/**
- * Set new FFT data.
  * @param fftData Pointer to the new FFT data used on the pandapter.
  * @param wfData Pointer to the FFT data used in the waterfall.
  * @param size The FFT size.
@@ -1856,7 +1011,7 @@ GLWaterfall::setNewFftData(
  * waterfall.
  */
 
-void
+  void
 GLWaterfall::setNewFftData(
     float *fftData,
     float *wfData,
@@ -1875,13 +1030,13 @@ GLWaterfall::setNewFftData(
 
     ts.counter = m_TimeStampCounter;
     ts.timeStampText =
-        m_lastFft.toLocalTime().toString("hh:mm:ss.zzz")
-          + " - "
-          + t.toLocalTime().toString("hh:mm:ss.zzz");
+      m_lastFft.toLocalTime().toString("hh:mm:ss.zzz")
+      + " - "
+      + t.toLocalTime().toString("hh:mm:ss.zzz");
     ts.utcTimeStampText =
-        m_lastFft.toUTC().toString("hh:mm:ss.zzzZ")
-          + " - "
-          + t.toUTC().toString("hh:mm:ss.zzzZ");
+      m_lastFft.toUTC().toString("hh:mm:ss.zzzZ")
+      + " - "
+      + t.toUTC().toString("hh:mm:ss.zzzZ");
     ts.marker = true;
 
     m_TimeStamps.push_front(ts);
@@ -1891,8 +1046,12 @@ GLWaterfall::setNewFftData(
   m_wfData = wfData;
   m_fftData = fftData;
   m_fftDataSize = size;
-  m_tentativeCenterFreq = 0;
   m_lastFft = t;
+
+  if (m_tentativeCenterFreq != 0) {
+    m_tentativeCenterFreq = 0;
+    m_DrawOverlay = true;
+  }
 
   if (m_TimeStampCounter >= m_TimeStampSpacing) {
     TimeStamp ts;
@@ -1932,150 +1091,7 @@ GLWaterfall::setNewFftData(
   draw();
 }
 
-void
-GLWaterfall::getScreenIntegerFFTData(
-    qint32 plotHeight,
-    qint32 plotWidth,
-    float maxdB,
-    float mindB,
-    qint64 startFreq,
-    qint64 stopFreq,
-    float *inBuf,
-    qint32 *outBuf,
-    int *xmin,
-    int *xmax)
-{
-  qint32 i;
-  qint32 y;
-  qint32 x;
-  qint32 ymax = 10000;
-  qint32 xprev = -1;
-  qint32 minbin, maxbin;
-  qint32 m_BinMin, m_BinMax;
-  qint32 m_FFTSize = m_fftDataSize;
-  float *m_pFFTAveBuf = inBuf;
-
-  mindB -= m_gain;
-  maxdB -= m_gain;
-
-  float  dBGainFactor = ((float)plotHeight) / fabs(maxdB - mindB);
-  qint32* m_pTranslateTbl = new qint32[qMax(m_FFTSize, plotWidth)];
-
-  /** FIXME: qint64 -> qint32 **/
-  m_BinMin = (qint32)((float)startFreq * (float)m_FFTSize / m_SampleFreq);
-  m_BinMin += (m_FFTSize/2);
-  m_BinMax = (qint32)((float)stopFreq * (float)m_FFTSize / m_SampleFreq);
-  m_BinMax += (m_FFTSize/2);
-
-  minbin = qBound(0, m_BinMin, m_FFTSize - 1);
-  maxbin = qBound(0, m_BinMax, m_FFTSize - 1);
-
-  bool largeFft = (maxbin - minbin) > plotWidth; // true if more fft point than plot points
-
-  if (largeFft) {
-    // more FFT points than plot points
-    for (i = minbin; i < maxbin; i++)
-      m_pTranslateTbl[i] = ((qint64)(i-m_BinMin)*plotWidth) / (m_BinMax - m_BinMin);
-    *xmin = m_pTranslateTbl[minbin];
-    *xmax = m_pTranslateTbl[maxbin - 1];
-  } else {
-    // more plot points than FFT points
-    for (i = 0; i < plotWidth; i++)
-      m_pTranslateTbl[i] = m_BinMin + (i*(m_BinMax - m_BinMin)) / plotWidth;
-    *xmin = 0;
-    *xmax = plotWidth;
-  }
-
-  if (largeFft) {
-    // more FFT points than plot points
-    for (i = minbin; i < maxbin; i++) {
-      y = (qint32)(dBGainFactor*(maxdB-m_pFFTAveBuf[i]));
-
-      if (y > plotHeight)
-        y = plotHeight;
-      else if (y < 0)
-        y = 0;
-
-      x = m_pTranslateTbl[i];	//get fft bin to plot x coordinate transform
-
-      if (x == xprev) {
-        if (y < ymax) {
-          outBuf[x] = y;
-          ymax = y;
-        }
-      } else {
-        outBuf[x] = y;
-        xprev = x;
-        ymax = y;
-      }
-    }
-  } else {
-    // more plot points than FFT points
-    for (x = 0; x < plotWidth; x++) {
-      i = m_pTranslateTbl[x]; // get plot to fft bin coordinate transform
-      if(i < 0 || i >= m_FFTSize)
-        y = plotHeight;
-      else
-        y = (qint32)(dBGainFactor*(maxdB-m_pFFTAveBuf[i]));
-
-      if (y > plotHeight)
-        y = plotHeight;
-      else if (y < 0)
-        y = 0;
-
-      outBuf[x] = y;
-    }
-  }
-
-  delete [] m_pTranslateTbl;
-}
-
-void
-GLWaterfall::setFftRange(float min, float max)
-{
-  setWaterfallRange(min, max);
-  setPandapterRange(min, max);
-}
-
-void
-GLWaterfall::setPandapterRange(float min, float max)
-{
-  if (out_of_range(min, max))
-    return;
-
-  m_PandMindB = min;
-  m_PandMaxdB = max;
-  updateOverlay();
-  m_PeakHoldValid = false;
-}
-
-void
-GLWaterfall::setWaterfallRange(float min, float max)
-{
-  if (out_of_range(min, max))
-    return;
-
-  m_WfMindB = min;
-  m_WfMaxdB = max;
-
-  this->glCtx.setDynamicRange(min - m_gain, max - m_gain);
-
-  // no overlay change is necessary
-}
-
-void GLWaterfall::setInfoText(QString const &text)
-{
-  m_infoText = text;
-  updateOverlay();
-}
-
-void GLWaterfall::setInfoTextColor(QColor const &color)
-{
-  m_infoTextColor = color;
-  updateOverlay();
-}
-
-int
+  int
 GLWaterfall::drawFATs(
     GLDrawingContext &ctx,
     qint64 StartFreq,
@@ -2121,29 +1137,29 @@ GLWaterfall::drawFATs(
         ctx.painter->setPen(p->second.color);
 
         ctx.painter->drawRect(
-              x0,
-              count * ctx.metrics->height(),
-              x1 - x0 + 1,
-              ctx.metrics->height());
+            x0,
+            count * ctx.metrics->height(),
+            x1 - x0 + 1,
+            ctx.metrics->height());
 
         if (leftborder)
           ctx.painter->drawLine(
-                x0,
-                count * ctx.metrics->height(),
-                x0,
-                h);
+              x0,
+              count * ctx.metrics->height(),
+              x0,
+              h);
 
         if (rightborder)
-            ctx.painter->drawLine(
-                  x1,
-                  count * ctx.metrics->height(),
-                  x1,
-                  h);
+          ctx.painter->drawLine(
+              x1,
+              count * ctx.metrics->height(),
+              x1,
+              h);
 
         label = ctx.metrics->elidedText(
-              QString::fromStdString(p->second.primary),
-              Qt::ElideRight,
-              boxw);
+            QString::fromStdString(p->second.primary),
+            Qt::ElideRight,
+            boxw);
 
 #if QT_VERSION >= QT_VERSION_CHECK(5, 11, 0)
         tw = ctx.metrics->horizontalAdvance(label);
@@ -2154,10 +1170,10 @@ GLWaterfall::drawFATs(
         if (tw < boxw) {
           ctx.painter->setPen(m_FftTextColor);
           rect.setRect(
-                x0 + (x1 - x0) / 2 - tw / 2,
-                count * ctx.metrics->height(),
-                tw,
-                ctx.metrics->height());
+              x0 + (x1 - x0) / 2 - tw / 2,
+              count * ctx.metrics->height(),
+              tw,
+              ctx.metrics->height());
           ctx.painter->drawText(rect, Qt::AlignHCenter | Qt::AlignVCenter, label);
         }
       }
@@ -2169,7 +1185,7 @@ GLWaterfall::drawFATs(
   return count * ctx.metrics->height();
 }
 
-void
+  void
 GLWaterfall::drawBookmarks(
     GLDrawingContext &ctx,
     qint64 StartFreq,
@@ -2184,7 +1200,7 @@ GLWaterfall::drawBookmarks(
   const int nLevels = 10;
 
   QList<BookmarkInfo> bookmarks =
-      m_BookmarkSource->getBookmarksInRange(
+    m_BookmarkSource->getBookmarksInRange(
         StartFreq,
         EndFreq);
   int tagEnd[nLevels] = {0};
@@ -2208,56 +1224,56 @@ GLWaterfall::drawBookmarks(
     tagEnd[level] = x + nameWidth + slant - 1;
 
 #if QT_VERSION >= QT_VERSION_CHECK(6, 0, 0)
-            BookmarkInfo bookmark = bookmarks[i];
-            m_BookmarkTags.append(
-                  qMakePair<QRect, BookmarkInfo>(
-                    QRect(x, yMin + level * levelHeight, nameWidth + slant, fontHeight),
-                    std::move(bookmark))); // Be more Cobol every day
+    BookmarkInfo bookmark = bookmarks[i];
+    m_BookmarkTags.append(
+        qMakePair<QRect, BookmarkInfo>(
+          QRect(x, yMin + level * levelHeight, nameWidth + slant, fontHeight),
+          std::move(bookmark))); // Be more Cobol every day
 #else
-            m_BookmarkTags.append(
-                  qMakePair<QRect, BookmarkInfo>(
-                    QRect(x, yMin + level * levelHeight, nameWidth + slant, fontHeight),
-                    bookmarks[i]));
+    m_BookmarkTags.append(
+        qMakePair<QRect, BookmarkInfo>(
+          QRect(x, yMin + level * levelHeight, nameWidth + slant, fontHeight),
+          bookmarks[i]));
 #endif // QT_VERSION
 
     QColor color = QColor(bookmarks[i].color);
     color.setAlpha(0x60);
     // Vertical line
     ctx.painter->setPen(
-          QPen(color, 1, Qt::DashLine));
+        QPen(color, 1, Qt::DashLine));
     ctx.painter->drawLine(
-          x,
-          yMin + level * levelHeight + fontHeight + slant,
-          x,
-          xAxisTop);
+        x,
+        yMin + level * levelHeight + fontHeight + slant,
+        x,
+        xAxisTop);
 
     // Horizontal line
     ctx.painter->setPen(
-          QPen(color, 1, Qt::SolidLine));
+        QPen(color, 1, Qt::SolidLine));
     ctx.painter->drawLine(
-          x + slant, yMin + level * levelHeight + fontHeight,
-          x + nameWidth + slant - 1,
-          yMin + level * levelHeight + fontHeight);
+        x + slant, yMin + level * levelHeight + fontHeight,
+        x + nameWidth + slant - 1,
+        yMin + level * levelHeight + fontHeight);
     // Diagonal line
     ctx.painter->drawLine(
-          x + 1,
-          yMin + level * levelHeight + fontHeight + slant - 1,
-          x + slant - 1,
-          yMin + level * levelHeight + fontHeight + 1);
+        x + 1,
+        yMin + level * levelHeight + fontHeight + slant - 1,
+        x + slant - 1,
+        yMin + level * levelHeight + fontHeight + 1);
 
     color.setAlpha(0xFF);
     ctx.painter->setPen(QPen(color, 2, Qt::SolidLine));
     ctx.painter->drawText(
-          x + slant,
-          yMin + level * levelHeight,
-          nameWidth,
-          fontHeight,
-          Qt::AlignVCenter | Qt::AlignHCenter,
-          bookmarks[i].name);
+        x + slant,
+        yMin + level * levelHeight,
+        nameWidth,
+        fontHeight,
+        Qt::AlignVCenter | Qt::AlignHCenter,
+        bookmarks[i].name);
   }
 }
 
-void
+  void
 GLWaterfall::drawAxes(GLDrawingContext &ctx, qint64 StartFreq, qint64 EndFreq)
 {
   int w = ctx.width;
@@ -2281,7 +1297,7 @@ GLWaterfall::drawAxes(GLDrawingContext &ctx, qint64 StartFreq, qint64 EndFreq)
 #if QT_VERSION >= QT_VERSION_CHECK(5, 11, 0)
   int tw = ctx.metrics->horizontalAdvance("XXXX");
 #else
-   int tw = ctx.metrics->width("XXXX");
+  int tw = ctx.metrics->width("XXXX");
 #endif // QT_VERSION_CHECK
 
   m_YAxisWidth = tw + 2 * HOR_MARGIN;
@@ -2303,18 +1319,18 @@ GLWaterfall::drawAxes(GLDrawingContext &ctx, qint64 StartFreq, qint64 EndFreq)
   label.setNum(float(EndFreq / m_FreqUnits), 'f', m_FreqDigits);
 
 #if QT_VERSION >= QT_VERSION_CHECK(5, 11, 0)
-   tw = ctx.metrics->horizontalAdvance(label) + ctx.metrics->horizontalAdvance("O");
+  tw = ctx.metrics->horizontalAdvance(label) + ctx.metrics->horizontalAdvance("O");
 #else
-   tw = ctx.metrics->width(label) + ctx.metrics->width("O");
+  tw = ctx.metrics->width(label) + ctx.metrics->width("O");
 #endif // QT_VERSION_CHECK
 
   calcDivSize(
-        StartFreq,
-        EndFreq,
-        qMin(w / tw, HORZ_DIVS_MAX),
-        m_StartFreqAdj,
-        m_FreqPerDiv,
-        m_HorDivs);
+      StartFreq,
+      EndFreq,
+      qMin(w / tw, HORZ_DIVS_MAX),
+      m_StartFreqAdj,
+      m_FreqPerDiv,
+      m_HorDivs);
 
   pixperdiv = (float)w * (float) m_FreqPerDiv / (float) m_Span;
   adjoffset = pixperdiv * float (m_StartFreqAdj - StartFreq) / (float) m_FreqPerDiv;
@@ -2331,15 +1347,15 @@ GLWaterfall::drawAxes(GLDrawingContext &ctx, qint64 StartFreq, qint64 EndFreq)
   ctx.painter->setPen(m_FftTextColor);
   for (int i = 0; i <= m_HorDivs; i++) {
 #if QT_VERSION >= QT_VERSION_CHECK(5, 11, 0)
-   int tw = ctx.metrics->horizontalAdvance(m_HDivText[i]);
+    int tw = ctx.metrics->horizontalAdvance(m_HDivText[i]);
 #else
-   int tw = ctx.metrics->width(m_HDivText[i]);
+    int tw = ctx.metrics->width(m_HDivText[i]);
 #endif // QT_VERSION_CHECK
-   x = (int)((float)i*pixperdiv + adjoffset);
-   if (x > m_YAxisWidth) {
-     rect.setRect(x - tw/2, fLabelTop, tw, ctx.metrics->height());
-     ctx.painter->drawText(rect, Qt::AlignHCenter|Qt::AlignBottom, m_HDivText[i]);
-   }
+    x = (int)((float)i*pixperdiv + adjoffset);
+    if (x > m_YAxisWidth) {
+      rect.setRect(x - tw/2, fLabelTop, tw, ctx.metrics->height());
+      ctx.painter->drawText(rect, Qt::AlignHCenter|Qt::AlignBottom, m_HDivText[i]);
+    }
   }
 
 
@@ -2351,8 +1367,8 @@ GLWaterfall::drawAxes(GLDrawingContext &ctx, qint64 StartFreq, qint64 EndFreq)
   qint64 pandMaxUnit = unitSign * static_cast<qint64>(toDisplayUnits(m_PandMaxdB));
 
   calcDivSize(pandMinUnit, pandMaxUnit,
-                qMax(h/m_VdivDelta, VERT_DIVS_MIN), minUnitAdj64, unitDivSize,
-                m_VerDivs);
+      qMax(h/m_VdivDelta, VERT_DIVS_MIN), minUnitAdj64, unitDivSize,
+      m_VerDivs);
 
   unitStepSize = (float) unitDivSize;
   minUnitAdj = minUnitAdj64;
@@ -2362,8 +1378,8 @@ GLWaterfall::drawAxes(GLDrawingContext &ctx, qint64 StartFreq, qint64 EndFreq)
 
 #ifdef PLOTTER_DEBUG
   qDebug() << "minDb =" << m_PandMindB << "maxDb =" << m_PandMaxdB
-           << "mindbadj =" << mindbadj << "dbstepsize =" << dbstepsize
-           << "pixperdiv =" << pixperdiv << "adjoffset =" << adjoffset;
+    << "mindbadj =" << mindbadj << "dbstepsize =" << dbstepsize
+    << "pixperdiv =" << pixperdiv << "adjoffset =" << adjoffset;
 #endif
 
   ctx.painter->setPen(QPen(m_FftAxesColor, 1, Qt::DotLine));
@@ -2397,9 +1413,9 @@ GLWaterfall::drawAxes(GLDrawingContext &ctx, qint64 StartFreq, qint64 EndFreq)
       unit = minUnitAdj + unitStepSize * i;
       rect.setRect(HOR_MARGIN, y - th / 2, m_YAxisWidth, th);
       ctx.painter->drawText(
-            rect,
-            Qt::AlignRight|Qt::AlignVCenter,
-            QString::number(unitSign * unit));
+          rect,
+          Qt::AlignRight|Qt::AlignVCenter,
+          QString::number(unitSign * unit));
     }
   }
 
@@ -2415,11 +1431,12 @@ GLWaterfall::drawAxes(GLDrawingContext &ctx, qint64 StartFreq, qint64 EndFreq)
 
 // Called to draw an overlay bitmap containing grid and text that
 // does not need to be recreated every fft data update.
-void
+  void
 GLWaterfall::drawOverlay()
 {
   if (m_OverlayPixmap.isNull())
     return;
+
   int             bandY = 0;
   QFontMetrics    metrics(m_Font);
   QPainter        painter(&m_OverlayPixmap);
@@ -2459,30 +1476,30 @@ GLWaterfall::drawOverlay()
 
       if (p->bandLike) {
         WFHelpers::drawChannelBox(
-              painter,
-              ctx.height,
-              x_fMin,
-              x_fMax,
-              x_fCenter,
-              p->boxColor,
-              p->markerColor,
-              p->name,
-              p->markerColor,
-              ctx.metrics->height() / 2,
-              bandY + p->nestLevel * ctx.metrics->height());
+            painter,
+            ctx.height,
+            x_fMin,
+            x_fMax,
+            x_fCenter,
+            p->boxColor,
+            p->markerColor,
+            p->name,
+            p->markerColor,
+            ctx.metrics->height() / 2,
+            bandY + p->nestLevel * ctx.metrics->height());
       } else {
         WFHelpers::drawChannelBox(
-              painter,
-              ctx.height,
-              x_fMin,
-              x_fMax,
-              x_fCenter,
-              p->boxColor,
-              p->markerColor,
-              p->name,
-              QColor(),
-              -1,
-              bandY + p->nestLevel * ctx.metrics->height());
+            painter,
+            ctx.height,
+            x_fMin,
+            x_fMax,
+            x_fCenter,
+            p->boxColor,
+            p->markerColor,
+            p->name,
+            QColor(),
+            -1,
+            bandY + p->nestLevel * ctx.metrics->height());
       }
     }
   }
@@ -2495,9 +1512,9 @@ GLWaterfall::drawOverlay()
     pixRect.setWidth(pixRect.width() - 10);
 
     QRectF rect = painter.boundingRect(
-          pixRect,
-          flags,
-          m_infoText);
+        pixRect,
+        flags,
+        m_infoText);
 
     rect.setX(pixRect.width() - rect.width());
     rect.setY(0);
@@ -2516,466 +1533,4 @@ GLWaterfall::drawOverlay()
   }
 
   painter.end();
-}
-
-static QString
-formatFreqUnits(int units)
-{
-  switch (units) {
-    case 1:
-      return QString("");
-
-    case 1000:
-      return QString("K");
-
-    case 1000000:
-      return QString("M");
-
-    case 1000000000:
-      return QString("G");
-  }
-
-  return QString("");
-}
-
-
-// Create frequency division strings based on start frequency, span frequency,
-// and frequency units.
-// Places in QString array m_HDivText
-// Keeps all strings the same fractional length
-void GLWaterfall::makeFrequencyStrs()
-{
-    qint64  StartFreq = m_StartFreqAdj;
-    float   freq;
-    int     i,j, centerDiv;
-    bool    deltaRep;
-
-    if ((1 == m_FreqUnits) || (m_FreqDigits == 0))
-    {
-        // if units is Hz then just output integer freq
-        for (int i = 0; i <= m_HorDivs; i++)
-        {
-            freq = (float)StartFreq/(float)m_FreqUnits;
-            m_HDivText[i].setNum((int)freq);
-            StartFreq += m_FreqPerDiv;
-        }
-        return;
-    }
-    // here if is fractional frequency values
-    // so create max sized text based on frequency units
-    for (int i = 0; i <= m_HorDivs; i++)
-    {
-        freq = (float)StartFreq / (float)m_FreqUnits;
-        m_HDivText[i].setNum(freq,'f', m_FreqDigits);
-        StartFreq += m_FreqPerDiv;
-    }
-
-    // now find the division text with the longest non-zero digit
-    // to the right of the decimal point.
-    int max = 0;
-    for (i = 0; i <= m_HorDivs; i++)
-    {
-        int dp = m_HDivText[i].indexOf('.');
-        int l = m_HDivText[i].length()-1;
-        for (j = l; j > dp; j--)
-        {
-            if (m_HDivText[i][j] != '0')
-                break;
-        }
-        if ((j - dp) > max)
-            max = j - dp;
-    }
-
-    // Decide whether we should use the delta representation here. We identify
-    // this situation if log10(m_FreqPerDiv) - log10(m_FreqUnits) <= -m_FreqDigits.
-    // This is, if the division in axis units requires more decimals than
-    // the current representation, settle for delta representation
-
-    deltaRep = log10(m_FreqPerDiv) - log10(m_FreqUnits) <= -m_FreqDigits;
-
-    // truncate all strings to maximum fractional length
-    StartFreq  = m_StartFreqAdj;
-    centerDiv  = m_HorDivs / 2;
-
-    if (deltaRep) {
-      for (i = 0; i <= m_HorDivs; i++) {
-          if (i == centerDiv) {
-            m_HDivText[i] = SuWidgetsHelpers::formatQuantityFromDelta(
-                  StartFreq,
-                  m_FreqPerDiv,
-                  "Hz",
-                  false);
-          } else {
-            m_HDivText[i] = SuWidgetsHelpers::formatQuantityFromDelta(
-                  (i - centerDiv) * m_FreqPerDiv,
-                  m_FreqPerDiv,
-                  "Hz",
-                  true);
-          }
-
-          StartFreq += m_FreqPerDiv;
-      }
-    } else {
-      for (i = 0; i <= m_HorDivs; i++) {
-          freq = (float)StartFreq/(float)m_FreqUnits;
-          m_HDivText[i].setNum(freq,'f', max);
-          m_HDivText[i] += formatFreqUnits(m_FreqUnits);
-          StartFreq += m_FreqPerDiv;
-      }
-    }
-}
-
-// Convert from screen coordinate to frequency
-int GLWaterfall::xFromFreq(qint64 freq)
-{
-    int w = m_OverlayPixmap.width();
-    qint64 StartFreq = m_CenterFreq + m_FftCenter - m_Span/2;
-    int x = (int) w * ((qreal)freq - StartFreq)/(qreal)m_Span;
-    if (x < 0)
-        return 0;
-    if (x > (int)w)
-        return m_OverlayPixmap.width();
-    return x;
-}
-
-// Convert from screen coordinate to Fft frequency
-qint64 GLWaterfall::fftFreqFromX(int x)
-{
-    int w = width();
-    qreal f0 = m_FftCenter - m_Span / 2;
-    qint64 f = (qint64) (f0 + (qreal) m_Span * (qreal) x / (qreal) w);
-    return f;
-}
-
-// Convert from frequency to screen coordinate
-qint64 GLWaterfall::freqFromX(int x)
-{
-    int w = width();
-    qint64 StartFreq = m_CenterFreq + m_FftCenter - m_Span / 2;
-    qint64 f = (qint64)(StartFreq + (qreal)m_Span * (qreal)x / (qreal)w);
-    return f;
-}
-
-/** Calculate time offset of a given line on the waterfall */
-quint64 GLWaterfall::msecFromY(int y)
-{
-    // ensure we are in the waterfall region
-    if (y < m_OverlayPixmap.height())
-        return 0;
-
-    int dy = y - m_OverlayPixmap.height();
-
-    if (msec_per_wfline > 0)
-        return tlast_wf_ms - dy * msec_per_wfline;
-    else
-        return tlast_wf_ms - dy * 1000 / fft_rate;
-}
-
-// Round frequency to click resolution value
-qint64 GLWaterfall::roundFreq(qint64 freq, int resolution)
-{
-    qint64 delta = resolution;
-    qint64 delta_2 = delta / 2;
-    if (freq >= 0)
-        return (freq - (freq + delta_2) % delta + delta_2);
-    else
-        return (freq - (freq + delta_2) % delta - delta_2);
-}
-
-// Clamp demod freqeuency limits of m_DemodCenterFreq
-void GLWaterfall::clampDemodParameters()
-{
-    if(m_DemodLowCutFreq < m_FLowCmin)
-        m_DemodLowCutFreq = m_FLowCmin;
-    if(m_DemodLowCutFreq > m_FLowCmax)
-        m_DemodLowCutFreq = m_FLowCmax;
-
-    if(m_DemodHiCutFreq < m_FHiCmin)
-        m_DemodHiCutFreq = m_FHiCmin;
-    if(m_DemodHiCutFreq > m_FHiCmax)
-        m_DemodHiCutFreq = m_FHiCmax;
-}
-
-void GLWaterfall::setDemodRanges(qint64 FLowCmin, qint64 FLowCmax,
-                              qint64 FHiCmin, qint64 FHiCmax,
-                              bool symetric)
-{
-    m_FLowCmin=FLowCmin;
-    m_FLowCmax=FLowCmax;
-    m_FHiCmin=FHiCmin;
-    m_FHiCmax=FHiCmax;
-    m_symetric=symetric;
-    clampDemodParameters();
-    updateOverlay();
-}
-
-void GLWaterfall::setCenterFreq(qint64 f)
-{
-  f = boundCenterFreq(f);
-
-  if (m_CenterFreq == f)
-    return;
-
-  m_tentativeCenterFreq += f - m_CenterFreq;
-  m_CenterFreq = f;
-
-  updateOverlay();
-
-  m_PeakHoldValid = false;
-}
-
-void GLWaterfall::setFrequencyLimits(qint64 min, qint64 max)
-{
-  m_lowerFreqLimit = min;
-  m_upperFreqLimit = max;
-
-  if (m_enforceFreqLimits)
-    this->setCenterFreq(m_CenterFreq);
-}
-
-void GLWaterfall::setFrequencyLimitsEnabled(bool enabled)
-{
-  m_enforceFreqLimits = enabled;
-
-  if (m_enforceFreqLimits)
-    this->setCenterFreq(m_CenterFreq);
-}
-
-NamedChannelSetIterator GLWaterfall::addChannel(
-    QString name,
-    qint64 frequency,
-    qint32 fMin,
-    qint32 fMax,
-    QColor boxColor,
-    QColor markerColor,
-    QColor cutOffColor)
-{
-  auto it = m_channelSet.addChannel(
-        name,
-        frequency,
-        fMin,
-        fMax,
-        boxColor,
-        markerColor,
-        cutOffColor);
-
-  refreshChannel(it);
-
-  return it;
-}
-
-void GLWaterfall::removeChannel(NamedChannelSetIterator it)
-{
-  m_channelSet.remove(it);
-
-  updateOverlay();
-}
-
-void GLWaterfall::refreshChannel(NamedChannelSetIterator &it)
-{
-  if (m_channelSet.isOutOfPlace(it)) {
-    it = m_channelSet.relocate(it);
-  }
-
-  updateOverlay();
-}
-
-NamedChannelSetIterator GLWaterfall::findChannel(qint64 freq)
-{
-  return m_channelSet.find(freq);
-}
-
-NamedChannelSetIterator GLWaterfall::channelCBegin() const
-{
-  return m_channelSet.cbegin();
-}
-
-NamedChannelSetIterator GLWaterfall::channelCEnd() const
-{
-  return m_channelSet.cend();
-}
-
-// Ensure overlay is updated by either scheduling or forcing a redraw
-void GLWaterfall::updateOverlay()
-{
-  if (m_Running) {
-    m_DrawOverlay = true;
-    // If the update rate is slow, draw now.
-    if (this->slow())
-      draw();
-  } else {
-    // Not the case. Draw now.
-    drawOverlay();
-  }
-}
-
-/** Reset horizontal zoom to 100% and centered around 0. */
-void GLWaterfall::resetHorizontalZoom(void)
-{
-    setFftCenterFreq(0);
-    setSpanFreq(static_cast<qint64>(m_SampleFreq));
-    m_PeakHoldValid = false;
-    emit newZoomLevel(1);
-}
-
-/** Center FFT plot around 0 (corresponds to center freq). */
-void GLWaterfall::moveToCenterFreq(void)
-{
-    setFftCenterFreq(0);
-    updateOverlay();
-    m_PeakHoldValid = false;
-}
-
-/** Center FFT plot around the demodulator frequency. */
-void GLWaterfall::moveToDemodFreq(void)
-{
-    setFftCenterFreq(m_DemodCenterFreq-m_CenterFreq);
-    updateOverlay();
-
-    m_PeakHoldValid = false;
-}
-
-/** Set FFT plot color. */
-void GLWaterfall::setFftPlotColor(const QColor color)
-{
-    m_FftColor = color;
-    m_FftFillCol = color;
-    m_FftFillCol.setAlpha(0x1A);
-    m_PeakHoldColor = color;
-    m_PeakHoldColor.setAlpha(60);
-    updateOverlay();
-}
-
-/** Set filter box color */
-void GLWaterfall::setFilterBoxColor(const QColor color)
-{
-  m_FilterBoxColor = color;
-  updateOverlay();
-}
-
-/** Set timestamp color */
-void GLWaterfall::setTimeStampColor(const QColor color)
-{
-  m_TimeStampColor = color;
-  updateOverlay();
-}
-
-/** Set FFT bg color. */
-void GLWaterfall::setFftBgColor(const QColor color)
-{
-    m_FftBgColor = color;
-    updateOverlay();
-}
-
-/** Set FFT axes color. */
-void GLWaterfall::setFftAxesColor(const QColor color)
-{
-    m_FftCenterAxisColor = color;
-    m_FftAxesColor = color;
-}
-
-/** Set FFT text color. */
-void GLWaterfall::setFftTextColor(const QColor color)
-{
-    m_FftTextColor = color;
-    updateOverlay();
-}
-
-/** Enable/disable filling the area below the FFT plot. */
-void GLWaterfall::setFftFill(bool enabled)
-{
-    m_FftFill = enabled;
-}
-
-/** Set peak hold on or off. */
-void GLWaterfall::setPeakHold(bool enabled)
-{
-    m_PeakHoldActive = enabled;
-    m_PeakHoldValid = false;
-}
-
-/**
- * Set peak detection on or off.
- * @param enabled The new state of peak detection.
- * @param c Minimum distance of peaks from mean, in multiples of standard deviation.
- */
-void GLWaterfall::setPeakDetection(bool enabled, float c)
-{
-    if(!enabled || c <= 0)
-        m_PeakDetection = -1;
-    else
-        m_PeakDetection = c;
-}
-
-void GLWaterfall::calcDivSize (qint64 low, qint64 high, int divswanted, qint64 &adjlow, qint64 &step, int& divs)
-{
-#ifdef PLOTTER_DEBUG
-    qDebug() << "low: " << low;
-    qDebug() << "high: " << high;
-    qDebug() << "divswanted: " << divswanted;
-#endif
-
-    if (divswanted == 0)
-        return;
-
-    static const qint64 stepTable[] = { 1, 2, 5 };
-    static const int stepTableSize = sizeof (stepTable) / sizeof (stepTable[0]);
-    qint64 multiplier = 1;
-    step = 1;
-    qint64 divsLong = high - low;
-    int index = 0;
-    adjlow = (low / step) * step;
-
-    while (divsLong > divswanted)
-    {
-        step = stepTable[index] * multiplier;
-        divsLong = (high - low) / step;
-        adjlow = (low / step) * step;
-        index = index + 1;
-        if (index == stepTableSize)
-        {
-            index = 0;
-            multiplier = multiplier * 10;
-        }
-    }
-    if (adjlow < low)
-        adjlow += step;
-
-
-    divs = static_cast<int>(divsLong);
-
-#ifdef PLOTTER_DEBUG
-    qDebug() << "adjlow: "  << adjlow;
-    qDebug() << "step: " << step;
-    qDebug() << "divs: " << divs;
-#endif
-}
-
-void GLWaterfall::pushFAT(const FrequencyAllocationTable *fat)
-{
-  m_FATs[fat->getName()] = fat;
-
-  if (m_ShowFATs)
-    this->updateOverlay();
-}
-
-bool GLWaterfall::removeFAT(std::string const &name)
-{
-  auto p = m_FATs.find(name);
-
-  if (p == m_FATs.end())
-    return false;
-
-  m_FATs.erase(p);
-
-  if (m_ShowFATs)
-    this->updateOverlay();
-
-  return true;
-}
-
-void GLWaterfall::setFATsVisible(bool visible)
-{
-  m_ShowFATs = visible;
-  this->updateOverlay();
 }
