+++ abstractwaterfall.pri
@@ -1,0 +1,4 @@
+WIDGET_HEADERS += AbstractWaterfall.h
+
+HEADERS += AbstractWaterfall.h
+SOURCES += AbstractWaterfall.cpp
